% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{toHugo}
\alias{toHugo}
\title{toHugo}
\usage{
toHugo(gene, translatorTable = translatedGenes)
}
\arguments{
\item{gene}{Either a HUGO gene symbol string or a Entrez ID}
}
\value{
A HUGO symbol.
}
\description{
Converts a Entrez Id to HUGO using org.Hs.eg.db.
Returns identity if it is already a HUGO symbol.
}
\examples{
toHugo("MYC")
toHugo(4609)
toHugo("4609")
}
